requirements.txt
azure-ai-agents==1.2.0b6
azure-core==1.36.0
certifi==2025.11.12
charset-normalizer==3.4.4
idna==3.11
isodate==0.7.2
requests==2.32.5
typing_extensions==4.15.0
urllib3==2.5.0
python-dotenv
azure-identity
uvicorn